% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_methods.R
\name{add}
\alias{add}
\alias{add,ATO,ATO_det-method}
\alias{add,ATO,ATO_dep-method}
\alias{add,ATO,ATO_tag-method}
\alias{add,ATO,ATO_ani-method}
\alias{add,ATO,ATO_obs-method}
\title{Generic add function}
\usage{
add(x, value, append = FALSE, silent = FALSE)

\S4method{add}{ATO,ATO_det}(x, value, append = FALSE, silent = FALSE)

\S4method{add}{ATO,ATO_dep}(x, value, append = FALSE, silent = FALSE)

\S4method{add}{ATO,ATO_tag}(x, value, append = FALSE, silent = FALSE)

\S4method{add}{ATO,ATO_ani}(x, value, append = FALSE, silent = FALSE)

\S4method{add}{ATO,ATO_obs}(x, value, append = FALSE, silent = FALSE)
}
\arguments{
\item{x}{an ATO}

\item{value}{The data to be included.
The output of one of the make_*() functions.}

\item{append}{Logical: Should the new data be appended to the
data already present in the ATO?}

\item{silent}{Logical: Supress summary messages}
}
\value{
The updated ATO
}
\description{
Brings already-formatted ATO data into an ATO object.
Adequated methods are picked based on the type of data
being incorporated.
}
\examples{
# split away parts of the example ATO for the example
ani <- get_ani(example_ato)
dep <- get_dep(example_ato)

# add them to the ATO object using the add methods
x <- add(example_ato, ani)
x <- add(x, dep)

# clean up
rm(ani, dep, x)

}
\seealso{
\code{\link{make_ani}}, \code{\link{make_dep}},
\code{\link{make_det}}, \code{\link{make_tag}}, \code{\link{make_obs}}
}
