% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATO_class.R
\docType{class}
\name{ATO-class}
\alias{ATO-class}
\title{S4 class: ATO}
\description{
The Acoustic Telemetry Object (ATO) class contains slots that allow
researchers to package their study data in a standard format, to be used
by the R packages that collectively make the trackyverse.
}
\section{Slots}{

\describe{
\item{\code{det}}{The detections slot. See \code{\link{ATO_det}}.}

\item{\code{dep}}{The deployments slot. See \code{\link{ATO_dep}}.}

\item{\code{tag}}{The tags slot. See \code{\link{ATO_tag}}.}

\item{\code{ani}}{The animals slot. See \code{\link{ATO_ani}}.}

\item{\code{obs}}{The observations slot. See \code{\link{ATO_obs}}.}

\item{\code{log}}{The log slot. See \code{\link{ATO_log}}.}

\item{\code{tbl}}{The table style slot. See \code{\link{ATO_tbl}}.}

\item{\code{pkg}}{A list slot reserved for package outputs.}
}}

