% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATO_class.R
\docType{class}
\name{ATO_ani}
\alias{ATO_ani}
\alias{.ATO_ani}
\title{S4 class: ATO_ani}
\format{
See \code{\link{make_ani}}.

An object of class \code{ATO_ani} (inherits from \code{data.frame}) with 0 rows and 10 columns.
}
\usage{
.ATO_ani
}
\description{
An S4 class for the \code{\link{ATO}} animals (@ani).
Does not contain internal slots. It is a table of
either data.frame, data.table, or tibble format.
}
\details{
The prototype @ani slot
contains the standard columns of the @ani slot.
Other columns are allowed, but these are necessary,
and must be of the designated types.

The ani slot contains information on the animals tagged.

Can be of type data.frame, data.table, or tibble,
depending on the @tbl slot.
See \code{\link{table_type}} for more details on
\code{\link{ATO}} table types.
}
\seealso{
\code{\link{make_ani}}
}
\keyword{classes}
\keyword{datasets}
