% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{ato_clear_cache}
\alias{ato_clear_cache}
\title{Clear the ato cache}
\usage{
ato_clear_cache()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Deletes all locally cached files. The next call to any data
function will re-download.
}
\examples{
\donttest{
op <- options(ato.cache_dir = tempdir())
ato_clear_cache()
options(op)
}
}
\seealso{
Other configuration: 
\code{\link{ato_cache_info}()},
\code{\link{ato_meta}()}
}
\concept{configuration}
