% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_functions.R
\name{ato_debug}
\alias{ato_debug}
\title{Helper function to turn debug mode on or off}
\usage{
ato_debug(value = c(TRUE, FALSE))
}
\arguments{
\item{value}{logical. Should matching be done on the fly?}
}
\value{
the current value for ATO_match_immediate if no new value is provided
}
\description{
Helper function to turn debug mode on or off
}
\examples{
# Use without arguments to verify the current state, or
# with TRUE of FALSE to activate/deactivate debug mode
ato_debug()

}
