% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_functions.R
\name{ato_debug_header}
\alias{ato_debug_header}
\title{Helper function send a message at the start of each function}
\usage{
ato_debug_header()
}
\value{
nothing, called for side effects
}
\description{
This function is intended for developers. If you want to benefit
from the same debug properties as the ATO functions, include a call
to this function at the head of your own.
}
\examples{
# include this at the top of your function. e.g.
x <- function(arg) {
  ato_debug_header()
  return(arg * 2)
}

# and then use ato_debug to get a message when this function starts
old_ato_debug <- ato_debug()
ato_debug(TRUE)
x(arg = 2)
ato_debug(old_ato_debug)

}
