% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATO_class.R
\docType{class}
\name{ATO_dep}
\alias{ATO_dep}
\alias{.ATO_dep}
\title{S4 class: ATO_dep}
\format{
See \code{\link{make_dep}}.

An object of class \code{ATO_dep} (inherits from \code{data.frame}) with 0 rows and 18 columns.
}
\usage{
.ATO_dep
}
\description{
An S4 class for the \code{\link{ATO}} deployments (@dep).
Does not contain internal slots. It is a table of
either data.frame, data.table, or tibble format.
}
\details{
The prototype @dep slot
contains the standard columns of the @dep slot.
Other columns are allowed, but these are necessary,
and must be of the designated types.

Deployments include both the deployments of receivers (where the receiver
information is filled but the transmitter information isn't); transceivers
(where both the receiver and the transmitter information are filled in;
i.e. a receiver with a beacon tag); and reference tags (where there is no
receiver information, but there is transmitter information).

Can be of type data.frame, data.table, or tibble,
depending on the @tbl slot.
See \code{\link{table_type}} for more details on
\code{\link{ATO}} table types.
}
\seealso{
\code{\link{make_dep}}
}
\keyword{classes}
\keyword{datasets}
