% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{ato_match_immediate}
\alias{ato_match_immediate}
\title{Helper function to turn automatic matching on or off}
\usage{
ato_match_immediate(value = TRUE, silent = FALSE)
}
\arguments{
\item{value}{logical. Should matching be done on the fly?}

\item{silent}{Supresses messages}
}
\value{
the current value for ATO_match_immediate if no new value is provided
}
\description{
Matching of the slots' contents is essential to keep the ATO operational.
Do not turn automatic matching off unless you are working with extremely
large datasets that would cause heavy computation delays.
}
\examples{
old_match_immediate <- ato_match_immediate()
ato_match_immediate(FALSE)
x <- init_ato(ani = ani(example_ato),
             tag = tag(example_ato),
             det = det(example_ato),
             dep = dep(example_ato))
x <- match_update(x)

# clean up
ato_match_immediate(old_match_immediate)
rm(x, old_match_immediate)

}
