% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATO_class.R
\docType{class}
\name{ATO_obs}
\alias{ATO_obs}
\alias{.ATO_obs}
\title{S4 class: ATO_obs}
\format{
A data frame with 0 rows and 8 variables:
\describe{
\item{animal}{Name of the animal that was observed, character.}
\item{transmitter}{Transmitter code that was observed, character.}
\item{type}{Type of observation (e.g. directly seen, manual tracking), character.}
\item{terminal}{Was the animal permanently captured at the moment of observation? logical.}
\item{location}{Name of the place where the observation occurred, character.}
\item{datetime}{date and time of the observation, posixct format.}
\item{lat}{latitude of the observation. Preferably in WGS84, numeric.}
\item{lon}{longitude of the observation. Preferably in WGS84, numeric.}
}

An object of class \code{ATO_obs} (inherits from \code{data.frame}) with 0 rows and 9 columns.
}
\usage{
.ATO_obs
}
\description{
An S4 class for the \code{\link{ATO}} observations (@obs).
Does not contain internal slots. It is a table of
either data.frame, data.table, or tibble format.
}
\details{
The prototype @obs slot
contains the standard columns of the @obs slot.
Other columns are allowed, but these are necessary,
and must be of the designated types.

The observations slot contains information about locations where an animal
was seen or a tag was heard. E.g. fin observations, or detections through
manual tracking.

Can be of type data.frame, data.table, or tibble,
depending on the @tbl slot.
See \code{\link{table_type}} for more details on
\code{\link{ATO}} table types.
}
\seealso{
make_obs
}
\keyword{datasets}
\keyword{internal}
