% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATO_package.R
\docType{package}
\name{ATO_package}
\alias{ATO}
\alias{ATO-package}
\alias{ATO_package}
\title{ATO: Acoustic Telemetry Object Class}
\description{
Class and methods to create, edit, and expand an ATO.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trackyverse/ATO}
  \item \url{http://www.trackyverse.com/ATO}
  \item Report bugs at \url{https://github.com/trackyverse/ATO/issues}
}

}
\author{
\strong{Maintainer}: Hugo Flávio \email{hflavio@dal.ca}

Other contributors:
\itemize{
  \item Michael O'Brien \email{obrien@umces.edu} (\href{https://orcid.org/0000-0003-1420-6395}{ORCID}) [contributor]
}

}
\keyword{internal}
