% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_type_methods.R
\name{ato_table_type_global}
\alias{ato_table_type_global}
\title{Helper function to set the global type of ATO tables}
\usage{
ato_table_type_global(type = c("data.frame", "data.table", "tibble"))
}
\arguments{
\item{type}{the desired type of ATO tables. One of data.frame (default),
data.table (requires the package data.table installed), or
tibble (requires the package tibble installed).}
}
\value{
the current global table type if 'type' is missing.
}
\description{
Helper function to set the global type of ATO tables
}
\examples{
old_table_type_global <- ato_table_type_global()
ato_table_type_global("data.frame")

# clean up
ato_table_type_global(old_table_type_global)
rm(old_table_type_global)

}
