% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATO_class.R
\docType{class}
\name{ATO_tag}
\alias{ATO_tag}
\alias{.ATO_tag}
\title{S4 class: ATO_tag}
\format{
See \code{\link{make_tag}}.

An object of class \code{ATO_tag} (inherits from \code{data.frame}) with 0 rows and 13 columns.
}
\usage{
.ATO_tag
}
\description{
An S4 class for the \code{\link{ATO}} tags (@tag).
Does not contain internal slots. It is a table of
either data.frame, data.table, or tibble format.
}
\details{
The prototype @tag slot
contains the standard columns of the @tag slot.
Other columns are allowed, but these are necessary,
and must be of the designated types.

The tag slot contains information on the different transmitters being
tracked. Tags with multiple transmitter codes are coded as multiple rows
associated with a single serial number and a single animal.

Can be of type data.frame, data.table, or tibble,
depending on the @tbl slot.
See \code{\link{table_type}} for more details on
\code{\link{ATO}} table types.
}
\seealso{
\code{\link{make_tag}}
}
\keyword{classes}
\keyword{datasets}
