% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATO_class.R
\docType{class}
\name{ATO_tbl}
\alias{ATO_tbl}
\alias{.ATO_tbl}
\title{S4 class: ATO_tbl}
\format{
An object of class \code{ATO_tbl} (inherits from \code{character}) of length 1.
}
\usage{
.ATO_tbl
}
\description{
An S4 class for the \code{\link{ATO}} table format (@tbl).
Does not contain internal slots. It is a single value.
One of: "data.frame", "data.table", or "tibble"
}
\details{
The prototype @tbl slot contains the default value for the @log slot.
}
\keyword{classes}
\keyword{datasets}
