% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_ato_table_cols}
\alias{check_ato_table_cols}
\title{Check that the table columns match the prototype}
\usage{
check_ato_table_cols(object, ref)
}
\arguments{
\item{object}{A prospective ATO slot object}

\item{ref}{The prototype for the respective slot}
}
\value{
Nothing. Called to stop() if needed.
}
\description{
Check that the table columns match the prototype
}
\keyword{internal}
