% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_ato_table_type}
\alias{check_ato_table_type}
\title{Check that the table type matches the current global option}
\usage{
check_ato_table_type(object)
}
\arguments{
\item{object}{A prospective ATO slot object.}
}
\value{
Nothing. Called to stop() if needed.
}
\description{
Check that the table type matches the current global option
}
\keyword{internal}
