% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.check_ato_tzones}
\alias{.check_ato_tzones}
\title{Check that all slots inside the ATO are in the same timezone}
\usage{
.check_ato_tzones(x)
}
\arguments{
\item{x}{An ATO object}
}
\description{
Usage notes:
all in the same tz, returns the tz
\code{.check_ato_tzones(example_ato)}
}
\details{
force fail by different tz
\code{object <- example_ato}
\code{attributes(object@det$datetime)$tzone <- "UTC"}
\code{check_ato_tzones(object)}
}
\keyword{internal}
