% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.check_column_tzones}
\alias{.check_column_tzones}
\title{Check that all columns within a slot have the same timezone}
\usage{
.check_column_tzones(object, tz)
}
\arguments{
\item{object}{A prospective ATO slot table}

\item{tz}{optional; a timezone to apply to all time columns}
}
\description{
How to use:
\code{object <- get_tag(example_ato)}
}
\details{
all in the same tz, works
\code{object <- .check_column_tzones(object)}

force fail by missing tz
\code{attributes(object$release_datetime)$tzone <- ""}
\code{object <- .check_column_tzones(object)}

force fail by differing tzs
\code{attributes(object$release_datetime)$tzone <- "UTC"}
\code{object <- .check_column_tzones(object)}

correct fail by adding tz argument
\code{object <- .check_column_tzones(object, "Europe/Copenhagen")}

confirm correction applied
\code{attributes(object$activation_datetime)$tzone}
\code{attributes(object$release_datetime)$tzone}
}
\keyword{internal}
