% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.check_dup_match_base}
\alias{.check_dup_match_base}
\title{Check for duplicated matches in the base version}
\usage{
.check_dup_match_base(x, i, label_x, label_y)
}
\arguments{
\item{x}{the vector of a new subset of matches}

\item{i}{the tag in y currently being matched against x}

\item{label_x}{the name of the slot on the left side of the match}

\item{label_y}{the name of the slot on the right side of the match}
}
\value{
nothing, stops if a duplicate match is found.
}
\description{
This is called inside a for loop, where i is the
tag currently being matched to the detections
}
\keyword{internal}
