% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.check_dup_match_datatable}
\alias{.check_dup_match_datatable}
\title{Check for duplicated matches in the data.table version}
\usage{
.check_dup_match_datatable(x, y, original_x, original_y, label_x, label_y)
}
\arguments{
\item{x}{the vector of the xid from the data.table::foverlaps output}

\item{y}{the vector of the yid from the data.table::foverlaps output}

\item{original_x}{the original row order of x, for error messages}

\item{original_y}{the original row order of x, for error messages}

\item{label_x}{the name of the slot on the left side of the match}

\item{label_y}{the name of the slot on the right side of the match}
}
\value{
nothing, stops if duplicated match found
}
\description{
data.table can run all the matches in one go and still show
duplicated matches
}
\keyword{internal}
