% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.check_slot_tzones}
\alias{.check_slot_tzones}
\title{Check that all slots are in the same timezone}
\usage{
.check_slot_tzones(x)
}
\arguments{
\item{x}{An ATO object}
}
\description{
Check that all slots are in the same timezone
}
\keyword{internal}
