% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.check_tag_beacon}
\alias{.check_tag_beacon}
\title{Check the matches between tag and dep transmitters}
\usage{
.check_tag_beacon(x)
}
\arguments{
\item{x}{the ATO to be checked}
}
\value{
nothing. Stops if there's a fatal ambiguity.
}
\description{
Check the matches between tag and dep transmitters
}
\keyword{internal}
