% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.check_tbl_argument}
\alias{.check_tbl_argument}
\title{Check the tbl argument}
\usage{
.check_tbl_argument(tbl)
}
\arguments{
\item{tbl}{the tbl argument to be checked}
}
\description{
used by the make_ functions
}
\keyword{internal}
