% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.check_tibble_exists}
\alias{.check_tibble_exists}
\title{Check that tibble library is installed}
\usage{
.check_tibble_exists(error = TRUE)
}
\arguments{
\item{error}{Stop if not installed?}
}
\value{
TRUE if installed, FALSE if not installed and error = FALSE
}
\description{
Check that tibble library is installed
}
\keyword{internal}
