% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_functions.R
\name{.comma}
\alias{.comma}
\title{Concatenate vectors with commas}
\usage{
.comma(x, max_i = 5)
}
\arguments{
\item{x}{The vector to write out}

\item{max_i}{maximum number of vector elements to display before
saying "and n others"}
}
\value{
a string
}
\description{
Concatenate vectors with commas
}
\keyword{.comma(LETTERS[1:10])}
\keyword{.comma(LETTERS[1:10],}
\keyword{.comma(LETTERS[1:2])}
\keyword{.comma(LETTERS[1:2],}
\keyword{.comma(LETTERS[1:3])}
\keyword{.comma(LETTERS[1:3],}
\keyword{.comma(LETTERS[1:5])}
\keyword{.comma(LETTERS[1:5],}
\keyword{.comma(LETTERS[1])}
\keyword{.comma(LETTERS[1],}
\keyword{1)}
\keyword{2)}
\keyword{=}
\keyword{internal}
\keyword{max_i}
\keyword{tests:}
