% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_functions.R
\name{.data.table_exists}
\alias{.data.table_exists}
\title{Check that data.table library is installed}
\usage{
.data.table_exists(error = TRUE)
}
\arguments{
\item{error}{Stop if not installed?}
}
\value{
TRUE if installed, FALSE if not installed and error = FALSE
}
\description{
Check that data.table library is installed
}
\keyword{internal}
