% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_functions.R
\name{.es}
\alias{.es}
\title{Add es}
\usage{
.es(n, reverse = FALSE)
}
\arguments{
\item{n}{the number of elements}

\item{reverse}{logical, defaults to FALSE (see above)}
}
\value{
a string
}
\description{
Adds an “es” if n is > 1 and reverse is false (default)
Adds “es” if n is == 1 and reverse is true
}
\details{
e.g. match and matches
}
\keyword{internal}
