% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_functions.R
\name{.has}
\alias{.has}
\title{Choose between has and have}
\usage{
.has(n)
}
\arguments{
\item{n}{the number of elements}
}
\value{
a string
}
\description{
Choose between has and have
}
\keyword{internal}
