% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{.match_ani_obs}
\alias{.match_ani_obs}
\title{Match animals in the observations to the target animals}
\usage{
.match_ani_obs(x, silent = FALSE)
}
\arguments{
\item{x}{an \code{\link{ATO}}}

\item{silent}{Supresses summary messages}
}
\value{
the updated ATO
}
\description{
Automatically called by the \code{\link{set}} functions.
}
\keyword{internal}
