% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{.match_ani_tag}
\alias{.match_ani_tag}
\title{Match animals listed in the target tags to the target animals}
\usage{
.match_ani_tag(x, silent = FALSE)
}
\arguments{
\item{x}{an \code{\link{ATO}}}

\item{silent}{Supresses summary messages}
}
\value{
the updated ATO
}
\description{
Automatically called by the \code{\link{add}} functions.
}
\keyword{internal}
