% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{.match_dep_det_datatable}
\alias{.match_dep_det_datatable}
\title{Match transmitters in the detections to
beacon/reference transmitters in the deployments}
\usage{
.match_dep_det_datatable(x, silent)
}
\arguments{
\item{x}{an \code{\link{ATO}}}

\item{silent}{Supresses summary messages}
}
\value{
the updated ATO
}
\description{
Much faster than \code{\link{.match_dep_det_base}}
}
\details{
Automatically called by the \code{\link{match_update}} function.
}
\keyword{internal}
