% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{.match_tag_det}
\alias{.match_tag_det}
\title{Match transmitters in the detections to the target tags}
\usage{
.match_tag_det(x, silent = FALSE)
}
\arguments{
\item{x}{an \code{\link{ATO}}}

\item{silent}{Supresses summary messages}
}
\value{
the updated ATO
}
\description{
Automatically called by the \code{\link{add}} functions.
}
\keyword{internal}
