% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_functions.R
\name{.message_n_zero}
\alias{.message_n_zero}
\title{Count number of zeros}
\usage{
.message_n_zero(x, longslot, what, silent)
}
\arguments{
\item{x}{the vector of numbers}

\item{longslot}{long name of the slot (e.g. animals)}

\item{what}{what's being counted?}

\item{silent}{should a message be issued?}
}
\value{
The number of zeros.
}
\description{
Used by the match functions
}
\keyword{internal}
