% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_functions.R
\name{.message_orphan_dets}
\alias{.message_orphan_dets}
\title{Count orphan detections and issue message}
\usage{
.message_orphan_dets(x, silent = FALSE)
}
\arguments{
\item{x}{an ATO object}

\item{silent}{should a message be issued?}
}
\value{
The number of orphans.
}
\description{
Used by the match functions
}
\keyword{internal}
