% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_functions.R
\name{.message_stray_dets}
\alias{.message_stray_dets}
\title{Count stray detections and issue message}
\usage{
.message_stray_dets(x, silent = FALSE)
}
\arguments{
\item{x}{an ATO object}

\item{silent}{should a message be issued?}
}
\value{
The number of stray detections.
}
\description{
Used by the match functions
}
\keyword{internal}
