% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_functions.R
\name{.s}
\alias{.s}
\title{Add s}
\usage{
.s(n, reverse = FALSE)
}
\arguments{
\item{n}{the number of elements}

\item{reverse}{logical, defaults to FALSE (see above)}
}
\value{
a string
}
\description{
Adds an “s” if n is > 1 and reverse is false (default)
Adds “s” if n is == 1 and reverse is true
}
\details{
e.g. detection and detections
}
\keyword{internal}
