% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_functions.R
\name{.was}
\alias{.was}
\title{Choose between was and were}
\usage{
.was(n)
}
\arguments{
\item{n}{the number of elements}
}
\value{
a string
}
\description{
Choose between was and were
}
\keyword{internal}
