% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_ato.R
\name{example_ato}
\alias{example_ato}
\title{Example ATO object}
\format{
An S4 object with 8 slots:
\describe{
\item{ani}{Information about tagged animals. See \code{\link{make_ani}}}
\item{dep}{Information about deployments. See \code{\link{make_dep}}}
\item{det}{Information about detections. See \code{\link{make_det}}}
\item{tag}{Information about tags. See \code{\link{make_tag}}}
\item{obs}{Information about observations. See \code{\link{make_obs}}}
\item{log}{Log of actions. See \code{\link{event}}}
\item{tbl}{ATO table type. See \code{\link{table_type}}}
\item{pkg}{Reserved for other packages of the trackyverse}
}
}
\source{
The data in the example is the same as provided by the package actel,
and was collected by the author
}
\description{
The ATO package comes with a small example ATO object to allow for quick
exploration of the object's structure
}
\keyword{internal}
