% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_det}
\alias{get_det}
\alias{get_det,ATO-method}
\alias{det}
\title{Generic to extract detections as a table}
\usage{
get_det(x, receivers, transmitters, type = c("all", "valid", "invalid"))

\S4method{get_det}{ATO}(x, receivers, transmitters, type = c("all", "valid", "invalid"))

det(x, receivers, transmitters, type = c("all", "valid", "invalid"))
}
\arguments{
\item{x}{an \code{\link{ATO}} object}

\item{receivers}{An optional vector of receiver serial numbers from which
to extract detections.}

\item{transmitters}{An optional vector of transmitters for which to extract
detections.}

\item{type}{the type of rows to return. One of:
'all' - returns both valid and invalid rows (default);
'valid' - returns only valid rows;
'invalid' - returns only invalid rows}
}
\value{
a table of detections
}
\description{
Generic to extract detections as a table
}
\examples{
# extract all the detections from an ATO in table format
x <- get_det(example_ato)
summary(x)

# short form:
x <- det(example_ato)
summary(x)

# extract only detections from one or more specific receivers
x <- get_det(example_ato, receivers = "132908")
summary(x)

# or matching one or more specific transmitters
x <- get_det(example_ato, transmitters = "R64K-4529")
summary(x)

# or both!
x <- get_det(example_ato,
             receivers = "132908",
             transmitters = "R64K-4529")
summary(x)

# clean up
rm(x)

}
