% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_det_dep}
\alias{get_det_dep}
\alias{get_det_dep,ATO-method}
\title{A wrapper of \code{\link{get_det}} to extract detections for transmitters listed
in the dep slot}
\usage{
get_det_dep(x, receivers, type = c("all", "valid", "invalid"))

\S4method{get_det_dep}{ATO}(x, receivers, type = c("all", "valid", "invalid"))
}
\arguments{
\item{x}{an \code{\link{ATO}} object}

\item{receivers}{An optional vector of receiver serial numbers from which
to extract detections.}

\item{type}{the type of rows to return. One of:
'all' - returns both valid and invalid rows (default);
'valid' - returns only valid rows;
'invalid' - returns only invalid rows}
}
\value{
a table of detections
}
\description{
A wrapper of \code{\link{get_det}} to extract detections for transmitters listed
in the dep slot
}
\examples{
# wrapper to extract all detections that match transmitters
# listed in the @dep slot (beacon tags)
det <- get_det_dep(example_ato)
summary(det)
# note: The example_ato does not have beacon detections,
# so this returns a table with 0 rows

# extract only detections from one or more specific receivers
det <- get_det_dep(example_ato, receivers = "132908")
summary(det)
# note: The example_ato does not have beacon detections,
# so this returns a table with 0 rows

# clean up
rm(det)

}
