% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_det_tag}
\alias{get_det_tag}
\alias{get_det_tag,ATO-method}
\title{A wrapper of \code{\link{get_det}} to extract detections for transmitters
listed in the tag slot}
\usage{
get_det_tag(x, receivers, type = c("all", "valid", "invalid"))

\S4method{get_det_tag}{ATO}(x, receivers, type = c("all", "valid", "invalid"))
}
\arguments{
\item{x}{an \code{\link{ATO}} object}

\item{receivers}{An optional vector of receiver serial numbers from which
to extract detections.}

\item{type}{the type of rows to return. One of:
'all' - returns both valid and invalid rows (default);
'valid' - returns only valid rows;
'invalid' - returns only invalid rows}
}
\value{
a table of detections
}
\description{
A wrapper of \code{\link{get_det}} to extract detections for transmitters
listed in the tag slot
}
\examples{
# wrapper to extract all detections that match transmitters
# listed in the @tag slot
det <- get_det_tag(example_ato)

# extract only detections from one or more specific receivers
det <- get_det_tag(example_ato, receivers = "132908")
summary(det)

# clean up
rm(det)

}
