% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_log}
\alias{get_log}
\alias{get_log,ATO-method}
\title{Generic to extract the log slot as a table}
\usage{
get_log(x, debug = FALSE)

\S4method{get_log}{ATO}(x, debug = FALSE)
}
\arguments{
\item{x}{an \code{\link{ATO}} object}

\item{debug}{should debug entries and the call column be displayed?}
}
\value{
The log slot as a table
}
\description{
Generic to extract the log slot as a table
}
\examples{
# extract all the log lines from an ATO in table format
log <- get_log(example_ato)
head(log)

# clean up
rm(log)

}
