% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_methods.R
\name{has}
\alias{has}
\alias{has,ATO-method}
\title{Generic to check if an ATO has data in a given slot}
\usage{
has(object, value, column, allow_NA = TRUE, error = FALSE)

\S4method{has}{ATO}(object, value, column, allow_NA = TRUE, error = FALSE)
}
\arguments{
\item{object}{an \code{\link{ATO}}}

\item{value}{a vector with the names of the requested slots}

\item{column}{a vector of column names to find within the requested slots}

\item{allow_NA}{Should NAs be allowed in the inspected columns?}

\item{error}{Should the code execution stop if the requested slots are empty?}
}
\value{
TRUE if the slots have data, FALSE if they're empty and
error = FALSE. Throws an error if the slots are empty and error = TRUE.
}
\description{
Generic to check if an ATO has data in a given slot
}
\examples{
# check if ATO object has detections
has(example_ato, "det")

# check if ATO has observations
has(example_ato, "obs")

# alternatively, this would error
# has(example_ato, "obs", error = TRUE)

# check specific columns in slot
has(example_ato, "tag", c("release_datetime", "power_level"))
has(example_ato, "tag", c("test", "power_level"))
# has(example_ato, "tag", c("test", "power_level"), error = TRUE)

# check that columns don't have NAs
# has(example_ato, "tag",
#    c("release_datetime", "power_level"),
#    allow_NA = FALSE, error = TRUE)

}
