% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_functions.R
\name{init_ato}
\alias{init_ato}
\title{Initiate an \code{\link{ATO}}}
\usage{
init_ato(det, dep, tag, ani, obs)
}
\arguments{
\item{det}{an object of class ATO_det. See \code{\link{make_det}}.}

\item{dep}{an object of class ATO_dep. See \code{\link{make_dep}}.}

\item{tag}{an object of class ATO_tag. See \code{\link{make_tag}}.}

\item{ani}{an object of class ATO_ani. See \code{\link{make_ani}}.}

\item{obs}{an object of class ATO_obs. See \code{\link{make_obs}}.}
}
\value{
an \code{\link{ATO}}
}
\description{
This is a wrapper around the usual way of creating S4 objects through new().
The main purpose of this function is to ensure the ATO is created with the
right table type. See \code{\link{table_type}} for more details.
}
