% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_ato}
\alias{is_ato}
\title{Wrapper for is(x, "ATO")}
\usage{
is_ato(x, error = TRUE)
}
\arguments{
\item{x}{the object to test}

\item{error}{Should the code error if x is not an ATO?}
}
\value{
TRUE if x is an ATO, FALSE if it is not an ATO and error = FALSE.
Throws an error if x is not an ATO and error = TRUE.
}
\description{
This function allows an easy check if the input is an \code{\link{ATO}} and
throw an error if that is not the case.
}
