% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_comment}
\alias{log_comment}
\title{Log a new comment}
\usage{
log_comment(ato, ...)
}
\arguments{
\item{ato}{the ATO object to which to log the comment}

\item{...}{The text fragments that compose the comment.}
}
\value{
the updated ATO.
}
\description{
Logs a user-written comment into the log.
}
\examples{
# add an event to the ato
x <- log_comment(example_ato, "M: just a comment")
get_log(x)

}
