% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_event}
\alias{log_event}
\title{Log a new event}
\usage{
log_event(
  ato,
  type = c("log", "message", "msg", "warning", "debug", "comment"),
  ...,
  blame
)
}
\arguments{
\item{ato}{the ATO object to which to log the entry.}

\item{type}{One of several event flags:
\itemize{
\item 'log' saves to the log without displaying.
\item 'message' (or 'msg') displays a message on screen.
\item 'warning' displays a warning on screen.
\item 'debug' flags the entry as debug (not shown by default).
\item 'comment' user comment entry.
}}

\item{...}{The text fragments that compose the event message.}

\item{blame}{Not used yet; functionality to come in the future.}
}
\value{
The updated ATO.
}
\description{
Logs useful information as the analysis progresses. This function is meant to
be used by other functions within the trackyverse. If you are looking for a
way to add comments to the analysis, look at \code{\link{log_comment}}.
}
\examples{
# add an event to the ato
x <- log_event(example_ato, type = "msg", "M: just a test")
get_log(x)

}
