% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_functions.R
\name{make_ani}
\alias{make_ani}
\title{Make an ATO animals object}
\usage{
make_ani(
  animal,
  release_datetime,
  release_location = NA_character_,
  release_lat = NA_real_,
  release_lon = NA_real_,
  capture_datetime = as.POSIXct(NA_real_),
  capture_location = NA_character_,
  capture_lat = NA_real_,
  capture_lon = NA_real_,
  tz,
  tbl,
  ...
)
}
\arguments{
\item{animal}{Name of the animal being tracked (character). Required.}

\item{release_datetime}{date and time of the release (POSIXct). Required.}

\item{release_location}{Name of the location where the animal was released
character).}

\item{release_lat}{Latitude of the release. Preferably in WGS84 (numeric).}

\item{release_lon}{Longitude of the release. Preferably in WGS84 (numeric).}

\item{capture_datetime}{Date and time of the capture (POSIXct).}

\item{capture_location}{Name of the location where the animal was captured
(character).}

\item{capture_lat}{Latitude of the capture. Preferably in WGS84 (numeric).}

\item{capture_lon}{Longitude of the capture. Preferably in WGS84 (numeric).}

\item{tz}{the timezone of the datetime data.}

\item{tbl}{The type of table to be generated. One of "data.frame",
"data.table", or "tibble". If omitted, the output of
\code{\link{ato_table_type_global}} is used.}

\item{...}{Non-standard columns to be added to the table.}
}
\value{
an ATO_ani object, ready to be used by one of the \code{\link{set}}
functions or \code{\link{init_ato}}.
}
\description{
Formats the input data into the ATO format and appends the ATO_ani class.
}
\examples{
# only two columns are required to start an ani table.
# Note that posix objects created on-the-fly do not necessarily
# have an associated timezone - try running the code below 
# without the tz argument and it will fail.
x <- make_ani(animal = letters[1:10],
              release_datetime = rep(Sys.time(), 10),
              tz = "America/Halifax")
head(x)

# additionally, you can change the type of table being created by using 
# the tbl argument - this requires having either data.table or tibble
# installed on your machine.

}
\seealso{
\code{\link{ATO_ani}}
}
