% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_functions.R
\name{make_dep}
\alias{make_dep}
\title{Make an ATO deployments object}
\usage{
make_dep(
  deploy_datetime,
  recover_datetime,
  deploy_location = NA_character_,
  deploy_lat = NA_real_,
  deploy_lon = NA_real_,
  deploy_z = NA_real_,
  recover_lat = NA_real_,
  recover_lon = NA_real_,
  receiver_manufacturer = NA_character_,
  receiver_model = NA_character_,
  receiver_serial = NA_character_,
  receiver_codeset = NA_character_,
  transmitter = NA_character_,
  transmitter_manufacturer = NA_character_,
  transmitter_model = NA_character_,
  transmitter_serial = NA_character_,
  transmitter_ping_rate = NA_real_,
  tz,
  tbl,
  ...
)
}
\arguments{
\item{deploy_datetime}{date and time of the deployment (POSIXct). Required.}

\item{recover_datetime}{date and time of the recovery (POSIXct). Required.}

\item{deploy_location}{Name of the location where the receiver was deployed
(character).}

\item{deploy_lat}{latitude of the deployment. Preferably in WGS84 (numeric).}

\item{deploy_lon}{longitude of the deployment. Preferably in WGS84 (numeric).}

\item{deploy_z}{depth of the deployment, as measured from the reference
surface of the water body (numeric).}

\item{recover_lat}{latitude of the recovery point. Preferably in WGS84
(numeric).}

\item{recover_lon}{longitude of the recovery point. Preferably in WGS84
(numeric).}

\item{receiver_manufacturer}{Maker of the receiver (character).}

\item{receiver_model}{Model of the receiver (character).}

\item{receiver_serial}{Receiver serial number (character).}

\item{receiver_codeset}{Codeset of the receiver (character).}

\item{transmitter}{Transmitter code for a beacon/reference tag (character).}

\item{transmitter_manufacturer}{Manufacturer of the transmitter (character).}

\item{transmitter_model}{Model of the transmitter (character).}

\item{transmitter_serial}{Serial number of the transmitter (integer).}

\item{transmitter_ping_rate}{Expected ping rate of the transmitter,
in seconds (numeric).}

\item{tz}{the timezone of the datetime data.}

\item{tbl}{The type of table to be generated. One of "data.frame",
"data.table", or "tibble". If omitted, the output of
\code{\link{ato_table_type_global}} is used.}

\item{...}{Non-standard columns to be added to the table.}
}
\value{
an ATO_dep object, ready to be used by one of the \code{\link{set}}
functions or \code{\link{init_ato}}.
}
\description{
Formats the input data into the ATO format and appends the ATO_dep class.
}
\examples{
# only two columns are required to start a dep table.
# Note that posix objects created on-the-fly do not necessarily
# have an associated timezone - try running the code below 
# without the tz argument and it will fail.
x <- make_dep(deploy_datetime = rep(Sys.time(), 10),
              recover_datetime = rep(Sys.time(), 10),
              tz = "America/Halifax")
head(x)

# additionally, you can change the type of table being created by using 
# the tbl argument - this requires having either data.table or tibble
# installed on your machine.

# while you can create deployments with only deployment and recovery times,
# they will likely not be very useful unless you also include a receiver
# serial number or an associated transmitter code.

}
\seealso{
\code{\link{ATO_dep}}
}
