% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_functions.R
\name{make_tag}
\alias{make_tag}
\title{Make an ATO tags object}
\usage{
make_tag(
  transmitter,
  manufacturer = NA_character_,
  model = NA_character_,
  serial = NA_character_,
  power_level = NA_real_,
  ping_rate = NA_real_,
  ping_variation = NA_real_,
  activation_datetime = as.POSIXct(NA_real_),
  battery_life = NA_real_,
  sensor_type = NA_character_,
  sensor_unit = NA_character_,
  animal = NA_character_,
  tz,
  tbl,
  ...
)
}
\arguments{
\item{transmitter}{Transmitter code (character). Required.}

\item{manufacturer}{Manufacturer of the transmitter (character).}

\item{model}{Model of the transmitter (character).}

\item{serial}{Serial number of the tag (integer).}

\item{power_level}{Power level of the transmitter (real).}

\item{ping_rate, ping_variation}{Expected average ping rate of the
transmitter and respective variation around the average, in seconds
(numeric). E.g. if a tag's ping interval may vary between
60 and 120s, then ping_rate = 90, and ping_variation = 30.}

\item{activation_datetime}{Date and time of the tag activation (POSIXct).}

\item{battery_life}{Expected battery duration of the tag, in days (numeric).}

\item{sensor_type}{Type of sensor data associated with the transmitter
(character).}

\item{sensor_unit}{Unit of the data associated with the transmitter
(character).}

\item{animal}{Name of the animal that received the tag (character).}

\item{tz}{the timezone of the datetime data.}

\item{tbl}{The type of table to be generated. One of "data.frame",
"data.table", or "tibble". If omitted, the output of
getOption("ATO_table_type", default = "data.frame") is used.}

\item{...}{Non-standard columns to be added to the table.}
}
\value{
an ATO_tag object, ready to be used by one of the \code{\link{set}}
functions or \code{\link{init_ato}}.
}
\description{
Formats the input data into the ATO format and appends the ATO_tag class.
}
\seealso{
\code{\link{ATO_tag}}
}
