% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{match_det_dep}
\alias{match_det_dep}
\title{Match transmitters in the detections to
beacon/reference transmitters in the deployments}
\usage{
match_det_dep(x, silent = FALSE)
}
\arguments{
\item{x}{an \code{\link{ATO}}}

\item{silent}{Supresses summary messages}
}
\value{
the updated ATO
}
\description{
Automatically called by the \code{\link{add}} functions.
}
