% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{match_obs_tag}
\alias{match_obs_tag}
\title{Match transmitters in the observations to the target tags}
\usage{
match_obs_tag(x, silent = FALSE)
}
\arguments{
\item{x}{an \code{\link{ATO}}}

\item{silent}{Supresses summary messages}
}
\value{
the updated ATO
}
\description{
Automatically called by the \code{\link{add}} functions.
}
