% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_methods.R
\name{set}
\alias{set}
\alias{set_det}
\alias{set_det,ATO,ATO_det-method}
\alias{set_dep}
\alias{set_dep,ATO,ATO_dep-method}
\alias{set_tag}
\alias{set_tag,ATO,ATO_tag-method}
\alias{set_ani}
\alias{set_ani,ATO,ATO_ani-method}
\alias{set_obs}
\alias{set_obs,ATO,ATO_obs-method}
\alias{.set_worker}
\title{Set an ATO slot}
\usage{
set_det(x, value, append = FALSE, silent = FALSE)

\S4method{set_det}{ATO,ATO_det}(x, value, append = FALSE, silent = FALSE)

set_dep(x, value, append = FALSE, silent = FALSE)

\S4method{set_dep}{ATO,ATO_dep}(x, value, append = FALSE, silent = FALSE)

set_tag(x, value, append = FALSE, silent = FALSE)

\S4method{set_tag}{ATO,ATO_tag}(x, value, append = FALSE, silent = FALSE)

set_ani(x, value, append = FALSE, silent = FALSE)

\S4method{set_ani}{ATO,ATO_ani}(x, value, append = FALSE, silent = FALSE)

set_obs(x, value, append = FALSE, silent = FALSE)

\S4method{set_obs}{ATO,ATO_obs}(x, value, append = FALSE, silent = FALSE)

.set_worker(x, value, append = FALSE, silent = FALSE)
}
\arguments{
\item{x}{an ATO}

\item{value}{The data to be included. Use one of the make functions to
prepare it. See \code{\link{make_ani}}, \code{\link{make_dep}},
\code{\link{make_det}}, \code{\link{make_tag}}, and \code{\link{make_obs}}.}

\item{append}{Logical: Should the new data be appended to the
data already present in the ATO?}

\item{silent}{Logical: Supress summary messages}
}
\value{
The updated ATO
}
\description{
The family of set functions (set_ani, set_dep, set_det, set_tag, and set_obs)
brings already-formatted ATO data into an ATO object.
}
\examples{
# split away parts of the example ATO for the example
ani <- get_ani(example_ato)
dep <- get_dep(example_ato)

# add them to the ATO object using the set methods
x <- set_ani(example_ato, ani)
x <- set_dep(x, dep)

# clean up
rm(ani, dep, x)

}
\seealso{
\code{\link{make_ani}}, \code{\link{make_dep}},
\code{\link{make_det}}, \code{\link{make_tag}}, \code{\link{make_obs}}
}
\keyword{internal}
