% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_methods.R
\name{show,ATO-method}
\alias{show,ATO-method}
\title{Method to show an \code{\link{ATO}}}
\usage{
\S4method{show}{ATO}(object)
}
\arguments{
\item{object}{an ATO}
}
\value{
Nothing. Called for side-effects
}
\description{
Method to show an \code{\link{ATO}}
}
