% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,ATO_ani-method}
\alias{summary,ATO_ani-method}
\title{Summary method for an ATO_ani object}
\usage{
\S4method{summary}{ATO_ani}(object)
}
\arguments{
\item{object}{an ATO_ani object}
}
\value{
Nothing. Prints a summary.
}
\description{
Summary method for an ATO_ani object
}
