% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,ATO_dep-method}
\alias{summary,ATO_dep-method}
\title{Summary method for an ATO_dep object}
\usage{
\S4method{summary}{ATO_dep}(object)
}
\arguments{
\item{object}{an ATO_dep object}
}
\value{
Nothing. Prints a summary.
}
\description{
Summary method for an ATO_dep object
}
