% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,ATO_tag-method}
\alias{summary,ATO_tag-method}
\title{Summary method for an ATO_tag object}
\usage{
\S4method{summary}{ATO_tag}(object)
}
\arguments{
\item{object}{an ATO_tag object}
}
\value{
Nothing. Prints a summary.
}
\description{
Summary method for an ATO_tag object
}
