% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_type_methods.R
\name{table_type<-}
\alias{table_type<-}
\alias{table_type<-,ATO-method}
\alias{table_type<-,ATO_ani-method}
\alias{table_type<-,ATO_dep-method}
\alias{table_type<-,ATO_det-method}
\alias{table_type<-,ATO_obs-method}
\alias{table_type<-,ATO_tag-method}
\title{Assign a new table type to an ATO or an ATO slot.}
\usage{
table_type(x) <- value

\S4method{table_type}{ATO}(x) <- value

\S4method{table_type}{ATO_ani}(x) <- value

\S4method{table_type}{ATO_dep}(x) <- value

\S4method{table_type}{ATO_det}(x) <- value

\S4method{table_type}{ATO_obs}(x) <- value

\S4method{table_type}{ATO_tag}(x) <- value
}
\arguments{
\item{x}{an \code{\link{ATO}}}

\item{value}{The new table type.}
}
\value{
Acts directly on x
}
\description{
Assign a new table type to an ATO or an ATO slot.
}
