% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzone_methods.R
\name{tzone<-}
\alias{tzone<-}
\alias{tzone<-,ATO-method}
\alias{tzone<-,ATO_ani-method}
\alias{tzone<-,ATO_det-method}
\alias{tzone<-,ATO_dep-method}
\alias{tzone<-,ATO_obs-method}
\alias{tzone<-,ATO_tag-method}
\title{Modify the time zone of an ATO object}
\usage{
tzone(x) <- value

\S4method{tzone}{ATO}(x) <- value

\S4method{tzone}{ATO_ani}(x) <- value

\S4method{tzone}{ATO_det}(x) <- value

\S4method{tzone}{ATO_dep}(x) <- value

\S4method{tzone}{ATO_obs}(x) <- value

\S4method{tzone}{ATO_tag}(x) <- value
}
\arguments{
\item{x}{an \code{\link{ATO}} or ATO slot object.}

\item{value}{the desired time zone}
}
\value{
Nothing, acts directly on x
}
\description{
Modify the time zone of an ATO object
}
\examples{
# check the timezone of the example ATO
tzone(example_ato)
summary(get_det(example_ato))

# change the tz to UTC
tzone(example_ato) <- "UTC"
tzone(example_ato)
summary(get_det(example_ato))

# clean_up
tzone(example_ato) <- "Europe/Copenhagen"

}
