% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_methods.R
\name{add}
\alias{add}
\alias{add,ATO,ATO_det-method}
\alias{add,ATO,ATO_dep-method}
\alias{add,ATO,ATO_tag-method}
\alias{add,ATO,ATO_ani-method}
\alias{add,ATO,ATO_obs-method}
\title{Generic add function}
\usage{
add(x, value, append = FALSE)

\S4method{add}{ATO,ATO_det}(x, value, append = FALSE)

\S4method{add}{ATO,ATO_dep}(x, value, append = FALSE)

\S4method{add}{ATO,ATO_tag}(x, value, append = FALSE)

\S4method{add}{ATO,ATO_ani}(x, value, append = FALSE)

\S4method{add}{ATO,ATO_obs}(x, value, append = FALSE)
}
\arguments{
\item{x}{an ATO}

\item{value}{The data to be included.
The output of one of the make_* functions.}

\item{append}{Logical: Should the new data add to the
data already present in the ATO?}
}
\value{
The updated ATO
}
\description{
Brings already-formatted data into the ATO.
Adequated methods are picked based on the type of data
being incorporated.
}
