% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ato-package.R
\docType{package}
\name{ato-package}
\alias{ato}
\alias{ato-package}
\title{ato: Australian Taxation Office Data}
\description{
Tidy R access to Australian Taxation Office ('ATO') Taxation
Statistics and related datasets via the 'data.gov.au' CKAN API.
}
\section{Main function families}{

\itemize{
\item \code{\link[=ato_catalog]{ato_catalog()}}, \code{\link[=ato_download]{ato_download()}}: discovery and generic
resource access.
\item \code{\link[=ato_individuals]{ato_individuals()}}, \code{\link[=ato_individuals_postcode]{ato_individuals_postcode()}},
\code{\link[=ato_individuals_occupation]{ato_individuals_occupation()}}: individual tax return items.
\item \code{\link[=ato_companies]{ato_companies()}}: company tax by industry and turnover.
\item \code{\link[=ato_super_funds]{ato_super_funds()}}: APRA-regulated super funds and SMSFs.
\item \code{\link[=ato_top_taxpayers]{ato_top_taxpayers()}}: Corporate Tax Transparency release.
\item \code{\link[=ato_gst]{ato_gst()}}, \code{\link[=ato_industry]{ato_industry()}}: GST and industry aggregates.
\item \code{\link[=ato_cache_info]{ato_cache_info()}}, \code{\link[=ato_clear_cache]{ato_clear_cache()}}: cache management.
}
}

\section{Data source}{

Taxation Statistics are published annually by the ATO on
\url{https://www.ato.gov.au/about-ato/research-and-statistics/} and
mirrored at \url{https://data.gov.au/data/organization/australiantaxationoffice}.
Most datasets are licensed under Creative Commons Attribution 2.5
Australia; Corporate Tax Transparency is CC BY 3.0 Australia.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/charlescoverdale/ato}
  \item Report bugs at \url{https://github.com/charlescoverdale/ato/issues}
}

}
\author{
\strong{Maintainer}: Charles Coverdale \email{charlesfcoverdale@gmail.com}

}
\concept{Australian tax}
\concept{income distribution}
\concept{taxation statistics}
\keyword{internal}
