% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATO_class.R
\docType{class}
\name{ATO_ani}
\alias{ATO_ani}
\alias{.ATO_ani}
\title{S4 class: ATO_ani}
\format{
A data frame with 0 rows and 9 variables:
\describe{
\item{animal}{Name of the animal that received the tag, character.}
\item{capture_location}{Name of the location where the animal was captured, character.}
\item{capture_datetime}{date and time of the capture, posixct.}
\item{capture_lat}{latitude of the capture. Preferably in WGS84, numeric.}
\item{capture_lon}{longitude of the capture. Preferably in WGS84, numeric.}
\item{release_location}{Name of the location where the animal was released, character.}
\item{release_datetime}{date and time of the release, posixct.}
\item{release_lat}{latitude of the release. Preferably in WGS84, numeric.}
\item{release_lon}{longitude of the release. Preferably in WGS84, numeric.}
}

An object of class \code{ATO_ani} (inherits from \code{data.frame}) with 0 rows and 9 columns.
}
\usage{
.ATO_ani
}
\description{
An S4 class for the \code{\link{ATO}} animals (@ani).
Does not contain internal slots. It is a table of
either data.frame, data.table, or tibble format.
}
\details{
The prototype @ani slot
contains the standard columns of the @ani slot.
Other columns are allowed, but these are necessary,
and must be of the designated types.

The ani slot contains information on the animals tagged.

Can be of type data.frame, data.table, or tibble,
depending on the @tbl slot.
See \code{\link{table_type}} for more details on
\code{\link{ATO}} table types.
}
\seealso{
make_ani
}
\keyword{datasets}
\keyword{internal}
