% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATO_class.R
\docType{class}
\name{ATO_dep}
\alias{ATO_dep}
\alias{.ATO_dep}
\title{S4 class: ATO_dep}
\format{
A data frame with 0 rows and 16 variables:
\describe{
\item{receiver_model}{Model of the receiver, character.}
\item{receiver_serial}{Receiver serial number, integer.}
\item{receiver_codeset}{Codeset of the receiver, character.}
\item{deploy_location}{Name of the location where the receiver was deployed, character.}
\item{deploy_datetime}{date and time of the deployment, posixct.}
\item{deploy_lat}{latitude of the deployment. Preferably in WGS84, numeric.}
\item{deploy_lon}{longitude of the deployment. Preferably in WGS84, numeric.}
\item{deploy_z}{depth of the deployment, as measured from the reference surface of the water body, numeric.}
\item{recover_datetime}{date and time of the recovery, posixct.}
\item{recover_lat}{latitude of the recovery point. Preferably in WGS84, numeric.}
\item{recover_lon}{longitude of the recovery point. Preferably in WGS84, numeric.}
\item{transmitter}{Transmitter code for a beacon/reference tag, character.}
\item{transmitter_manufacturer}{Manufacturer of the transmitter, character.}
\item{transmitter_ping_rate}{Expected ping rate of the transmitter, numeric. In seconds.}
\item{transmitter_model}{Model of the transmitter, character.}
\item{transmitter_serial}{Serial number of the transmitter, integer.}
}

An object of class \code{ATO_dep} (inherits from \code{data.frame}) with 0 rows and 16 columns.
}
\usage{
.ATO_dep
}
\description{
An S4 class for the \code{\link{ATO}} deployments (@dep).
Does not contain internal slots. It is a table of
either data.frame, data.table, or tibble format.
}
\details{
The prototype @dep slot
contains the standard columns of the @dep slot.
Other columns are allowed, but these are necessary,
and must be of the designated types.

Deployments include both the deployments of receivers (where the receiver
information is filled but the transmitter information isn't); transceivers
(where both the receiver and the transmitter information are filled in;
i.e. a receiver with a beacon tag); and reference tags (where there is no
receiver information, but there is transmitter information).

Can be of type data.frame, data.table, or tibble,
depending on the @tbl slot.
See \code{\link{table_type}} for more details on
\code{\link{ATO}} table types.
}
\seealso{
make_dep
}
\keyword{datasets}
\keyword{internal}
