% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATO_class.R
\docType{class}
\name{ATO_det}
\alias{ATO_det}
\alias{.ATO_det}
\title{S4 class: ATO_det}
\format{
A data frame with 0 rows and 5 variables:
\describe{
\item{datetime}{date and time, posixct format.}
\item{frac_second}{fractional second, numeric.}
\item{receiver_serial}{receiver serial number, integer.}
\item{transmitter}{transmitter code, character.}
\item{sensor_value}{reported sensor value, numeric.}
}

An object of class \code{ATO_det} (inherits from \code{data.frame}) with 0 rows and 6 columns.
}
\usage{
.ATO_det
}
\description{
An S4 class for the \code{\link{ATO}} detections (@det).
Does not contain internal slots. It is a table of
either data.frame, data.table, or tibble format.
}
\details{
The prototype @det slot contains the standard columns
of the @det slot. Other columns are allowed, but these
are necessary, and must be of the designated types.

Can be of type data.frame, data.table, or tibble,
depending on the @tbl slot.
See \code{\link{table_type}} for more details on
\code{\link{ATO}} table types.
}
\seealso{
make_det
}
\keyword{datasets}
\keyword{internal}
