% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{$,ATO-method}
\alias{$,ATO-method}
\title{\code{$} Method to extract ATO slots directly}
\usage{
\S4method{$}{ATO}(x, name)
}
\arguments{
\item{x}{an \code{\link{ATO}} object}

\item{name}{the name of the desired ATO slot}
}
\value{
The slot as defined in \code{name}, along with a warning to directly
extract slots in the future.
}
\description{
\code{$} Method to extract ATO slots directly
}
\seealso{
\link{Extract}
}
