% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_methods.R
\name{check}
\alias{check}
\alias{check,ATO_det-method}
\alias{check,ATO_dep-method}
\alias{check,ATO_tag-method}
\alias{check,ATO_ani-method}
\alias{check,ATO_obs-method}
\alias{check,ATO_tbl-method}
\title{check generic}
\usage{
check(object)

\S4method{check}{ATO_det}(object)

\S4method{check}{ATO_dep}(object)

\S4method{check}{ATO_tag}(object)

\S4method{check}{ATO_ani}(object)

\S4method{check}{ATO_obs}(object)

\S4method{check}{ATO_tbl}(object)
}
\arguments{
\item{object}{The prospective ATO slot object}
}
\value{
Nothing. Called to stop() if needed.
}
\description{
Checks a prospective ATO slot object to confirm that the
contents are as expected. Matches it with the respective prototype reference.
}
\keyword{internal}
