% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.check_column_tzones}
\alias{.check_column_tzones}
\title{Check that all columns within a slot have the same timezone}
\usage{
.check_column_tzones(object, tz)
}
\arguments{
\item{object}{A prospective ATO slot table}

\item{tz}{optional; a timezone to apply to all time columns}
}
\description{
Check that all columns within a slot have the same timezone
}
\keyword{internal}
