% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_ani}
\alias{get_ani}
\alias{get_ani,ATO-method}
\title{Generic to extract the ani slot as a table}
\usage{
get_ani(x)

\S4method{get_ani}{ATO}(x)
}
\arguments{
\item{x}{an \code{\link{ATO}} object}
}
\value{
The ani slot as a table
}
\description{
Generic to extract the ani slot as a table
}
