% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_dep}
\alias{get_dep}
\alias{get_dep,ATO-method}
\title{Generic to extract the dep slot as a table}
\usage{
get_dep(x, type = c("all", "valid", "invalid"))

\S4method{get_dep}{ATO}(x, type = c("all", "valid", "invalid"))
}
\arguments{
\item{x}{an \code{\link{ATO}} object}

\item{type}{the type of rows to return. One of:
'all' - returns both valid and invalid rows (default);
'valid' - returns only valid rows;
'invalid' - returns only invalid rows}
}
\value{
The dep slot as a table
}
\description{
Generic to extract the dep slot as a table
}
