% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_det}
\alias{get_det}
\alias{get_det,ATO-method}
\title{Generic to extract detections as a table}
\usage{
get_det(x, receivers, transmitters)

\S4method{get_det}{ATO}(x, receivers, transmitters)
}
\arguments{
\item{x}{an \code{\link{ATO}} object}

\item{receivers}{An optional vector of receiver serial numbers from which
to extract detections.}

\item{transmitters}{An optional vector of transmitters for which to extract
detections.}
}
\value{
a table of detections
}
\description{
Generic to extract detections as a table
}
