% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_det_tag}
\alias{get_det_tag}
\alias{get_det_tag,ATO-method}
\title{A wrapper of \code{\link{get_det}} to extract detections for transmitters listed
in the tag slot}
\usage{
get_det_tag(x, receivers)

\S4method{get_det_tag}{ATO}(x, receivers)
}
\arguments{
\item{x}{an \code{\link{ATO}} object}

\item{receivers}{An optional vector of receiver serial numbers from which
to extract detections.}
}
\value{
a table of detections
}
\description{
A wrapper of \code{\link{get_det}} to extract detections for transmitters listed
in the tag slot
}
