% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_obs}
\alias{get_obs}
\alias{get_obs,ATO-method}
\title{Generic to extract the obs slot as a table}
\usage{
get_obs(x)

\S4method{get_obs}{ATO}(x)
}
\arguments{
\item{x}{an \code{\link{ATO}} object}
}
\value{
The obs slot as a table
}
\description{
Generic to extract the obs slot as a table
}
