% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_tag}
\alias{get_tag}
\alias{get_tag,ATO-method}
\title{Generic to extract the tag slot as a table}
\usage{
get_tag(x, type = c("all", "valid", "invalid"))

\S4method{get_tag}{ATO}(x, type = c("all", "valid", "invalid"))
}
\arguments{
\item{x}{an \code{\link{ATO}} object}
}
\value{
The tag slot as a table
}
\description{
Generic to extract the tag slot as a table
}
