% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_event}
\alias{log_event}
\title{Log a new event}
\usage{
log_event(
  ato,
  type = c("log", "message", "msg", "warning", "debug", "comment"),
  ...,
  blame
)
}
\arguments{
\item{ato}{the ATO object to which to log the entry.}

\item{type}{One of several event flags:
\itemize{
\item 'log' saves to the log without displaying.
\item 'message' (or 'msg') displays a message on screen.
\item 'warning' displays a warning on screen.
\item 'debug' flags the entry as debug (not shown by default).
\item 'comment' user comment entry.
}}

\item{...}{The text fragments that compose the event message.}

\item{blame}{Not used currently}
}
\value{
The updated ATO.
}
\description{
Logs useful information as the analysis progresses.
}
