% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_functions.R
\name{make_ani}
\alias{make_ani}
\title{Make an ATO animals object}
\usage{
make_ani(
  animal,
  capture_location = NA_character_,
  capture_datetime = as.POSIXct(NA_real_),
  capture_lat = NA_real_,
  capture_lon = NA_real_,
  release_location,
  release_datetime,
  release_lat = NA_real_,
  release_lon = NA_real_,
  tz,
  ...
)
}
\arguments{
\item{animal}{Name of the animal that received the tag, character. Mandatory.}

\item{capture_location}{Name of the location where the animal was captured,
character. Optional.}

\item{capture_datetime}{date and time of the capture, posixct. Optional.}

\item{capture_lat}{latitude of the capture. Preferably in WGS84, numeric.
Optional.}

\item{capture_lon}{longitude of the capture. Preferably in WGS84, numeric.
Optional.}

\item{release_location}{Name of the location where the animal was released,
character. Mandatory.}

\item{release_datetime}{date and time of the release, posixct. Mandatory.}

\item{release_lat}{latitude of the release. Preferably in WGS84, numeric.
Optional.}

\item{release_lon}{longitude of the release. Preferably in WGS84, numeric.
Optional.}

\item{tz}{the timezone of the datetime data.}

\item{...}{Non-standard columns to be added to the table.}
}
\value{
an ATO_ani object, ready to be used by \code{\link{add}} or
\code{\link{init_ato}}.
}
\description{
Formats the input data into the ATO format and appends the ATO_ani class.
}
\seealso{
ATO_ani
}
