% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_functions.R
\name{make_dep}
\alias{make_dep}
\title{Make an ATO deployments object}
\usage{
make_dep(
  receiver_model = NA_character_,
  receiver_serial,
  receiver_codeset = NA_character_,
  deploy_location,
  deploy_datetime,
  deploy_lat = NA_real_,
  deploy_lon = NA_real_,
  deploy_z = NA_real_,
  recover_datetime,
  recover_lat = NA_real_,
  recover_lon = NA_real_,
  transmitter = NA_character_,
  transmitter_manufacturer = NA_character_,
  transmitter_ping_rate = NA_real_,
  transmitter_model = NA_character_,
  transmitter_serial = NA_character_,
  tz,
  ...
)
}
\arguments{
\item{receiver_model}{Model of the receiver, character. Optional.}

\item{receiver_serial}{Receiver serial number, integer. Mandatory.}

\item{receiver_codeset}{Codeset of the receiver, character. Optional.}

\item{deploy_location}{Name of the location where the receiver was deployed,
character. Mandatory.}

\item{deploy_datetime}{date and time of the deployment, posixct. Mandatory.}

\item{deploy_lat}{latitude of the deployment. Preferably in WGS84, numeric.
Optional.}

\item{deploy_lon}{longitude of the deployment. Preferably in WGS84, numeric.
Optional.}

\item{deploy_z}{depth of the deployment, as measured from the reference
surface of the water body, numeric. Optional.}

\item{recover_datetime}{date and time of the recovery, posixct. Mandatory.}

\item{recover_lat}{latitude of the recovery point. Preferably in WGS84,
numeric. Optional.}

\item{recover_lon}{longitude of the recovery point. Preferably in WGS84,
numeric. Optional.}

\item{transmitter}{Transmitter code for a beacon/reference tag, character.
Optional.}

\item{transmitter_manufacturer}{Manufacturer of the transmitter, character.
Optional.}

\item{transmitter_ping_rate}{Expected ping rate of the transmitter, numeric.
In seconds. Required if transmitter is provided.}

\item{transmitter_model}{Model of the transmitter, character. Optional.}

\item{transmitter_serial}{Serial number of the transmitter, integer.
Required if transmitter is provided.}

\item{tz}{the timezone of the datetime data. Mandatory.}

\item{...}{Non-standard columns to be added to the table.}
}
\value{
an ATO_dep object, ready to be used by \code{\link{add}} or
\code{\link{init_ato}}.
}
\description{
Formats the input data into the ATO format and appends the ATO_dep class.
}
\seealso{
ATO_dep
}
