% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_functions.R
\name{make_det}
\alias{make_det}
\title{Make an ATO detections object}
\usage{
make_det(
  datetime,
  receiver_serial,
  transmitter,
  frac_second = NA_real_,
  sensor_value = NA_real_,
  tz,
  tbl,
  ...
)
}
\arguments{
\item{datetime}{date and time of the detection (POSIXct). Required.}

\item{receiver_serial}{Receiver serial (character). Required.}

\item{transmitter}{Transmitter code (character). Required.}

\item{frac_second}{fractional second of the detection (numeric).}

\item{sensor_value}{reported sensor value (numeric).}

\item{tz}{the timezone of the datetime data.}

\item{tbl}{The type of table to be generated. One of "data.frame",
"data.table", or "tibble". If omitted, the output of
getOption("ATO_table_type", default = "data.frame") is used.}

\item{...}{Non-standard columns to be added to the table.}
}
\value{
an ATO_det object, ready to be used by one of the \code{\link{set}}
functions or \code{\link{init_ato}}.
}
\description{
Formats the input data into the ATO format and appends the ATO_det class.
}
\seealso{
\code{\link{ATO_det}}
}
