% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_functions.R
\name{make_det}
\alias{make_det}
\title{Make an ATO detections object}
\usage{
make_det(
  datetime,
  frac_second = NA_real_,
  receiver_serial,
  transmitter,
  sensor_value = NA_real_,
  tz,
  ...
)
}
\arguments{
\item{datetime}{date and time, posixct format. Mandatory.}

\item{frac_second}{fractional second, numeric. Optional.}

\item{receiver_serial}{Mandatory. receiver serial number, integer. Mandatory.}

\item{transmitter}{Mandatory. transmitter code, character. Mandatory.}

\item{sensor_value}{reported sensor value, numeric. Optional.}

\item{tz}{the timezone of the datetime data. Mandatory.}

\item{...}{Non-standard columns to be added to the table. Optional.}
}
\value{
an ATO_det object, ready to be used by \code{\link{add}} or
\code{\link{init_ato}}.
}
\description{
Formats the input data into the ATO format and appends the ATO_det class.
}
\seealso{
ATO_det
}
