% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_functions.R
\name{make_det}
\alias{make_det}
\title{Make an ATO detections object}
\usage{
make_det(
  datetime = as.POSIXct(NA_real_),
  frac_second = NA_real_,
  receiver_serial = NA_character_,
  transmitter = NA_character_,
  sensor_value = NA_real_,
  tz,
  ...
)
}
\arguments{
\item{datetime}{date and time, posixct format.}

\item{frac_second}{fractional second, numeric.}

\item{receiver_serial}{receiver serial number, integer.}

\item{transmitter}{transmitter code, character.}

\item{sensor_value}{reported sensor value, numeric.}

\item{tz}{the timezone of the datetime data.}

\item{...}{Non-standard columns to be added to the table.}
}
\value{
an ATO_det object, ready to be used by \code{\link{add}} or
\code{\link{init_ato}}.
}
\description{
Formats the input data into the ATO format and appends the ATO_det class.
}
\seealso{
ATO_det
}
