% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_functions.R
\name{make_obs}
\alias{make_obs}
\title{Make an ATO observations object}
\usage{
make_obs(
  datetime,
  animal = NA_character_,
  transmitter = NA_character_,
  terminal,
  type = NA_character_,
  location,
  lat = NA_real_,
  lon = NA_real_,
  tz,
  ...
)
}
\arguments{
\item{datetime}{date and time of the observation (POSIXct). Required.}

\item{animal}{Name of the animal that was observed (character). Each
observation must have animal or transmitter information (or both).}

\item{transmitter}{Transmitter code that was observed (character). Each
observation must have animal or transmitter information (or both).}

\item{terminal}{Was the animal/transmitter permanently captured at the
moment of observation (logical)? Required.}

\item{type}{Type of observation (e.g. directly seen, manual tracking)
(character).}

\item{location}{Name of the place where the observation occurred (character).}

\item{lat}{latitude of the observation. Preferably in WGS84 (numeric).}

\item{lon}{longitude of the observation. Preferably in WGS84 (numeric).}

\item{tz}{the timezone of the datetime data.}

\item{...}{Non-standard columns to be added to the table.}
}
\value{
an ATO_obs object, ready to be used by one of the \code{\link{set}}
functions or \code{\link{init_ato}}.
}
\description{
Formats the input data into the ATO format and appends the ATO_obs class.
}
\seealso{
\code{\link{ATO_obs}}
}
