% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_functions.R
\name{make_tag}
\alias{make_tag}
\title{Make an ATO tags object}
\usage{
make_tag(
  manufacturer = NA_character_,
  model = NA_character_,
  power_level = NA_real_,
  ping_rate = NA_real_,
  ping_variation = NA_real_,
  serial = NA_character_,
  transmitter = NA_character_,
  activation_datetime = as.POSIXct(NA_real_),
  battery_life = NA_real_,
  sensor_type = NA_character_,
  sensor_unit = NA_character_,
  animal = NA_character_,
  tz,
  ...
)
}
\arguments{
\item{manufacturer}{Manufacturer of the transmitter, character.}

\item{model}{Model of the transmitter, character.}

\item{power_level}{Power level of the transmitter, real.}

\item{ping_rate}{Expected ping rate of the transmitter, numeric. In seconds.}

\item{ping_variation}{Range of the variation added between pings,
to reduce tag collisions, numeric. In seconds}

\item{serial}{Serial number of the tag, integer.}

\item{transmitter}{Transmitter code, character.}

\item{activation_datetime}{date and time of the tag activation, posixct.}

\item{battery_life}{expected battery duration of the tag, numeric.}

\item{sensor_type}{Type of sensor data associated with the transmitter,
character.}

\item{sensor_unit}{Unit of the data associated with the transmitter,
character.}

\item{animal}{Name of the animal that received the tag, character.}

\item{tz}{the timezone of the datetime data.}

\item{...}{Non-standard columns to be added to the table.}
}
\value{
an ATO_tag object, ready to be used by \code{\link{add}} or
\code{\link{init_ato}}.
}
\description{
Formats the input data into the ATO format and appends the ATO_tag class.
}
\seealso{
ATO_tag
}
