% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{match_update}
\alias{match_update}
\title{Match the various slots of the ATO object}
\usage{
match_update(x, silent = FALSE)
}
\arguments{
\item{x}{an \code{\link{ATO}}}

\item{silent}{Supresses summary messages}
}
\value{
the updated ATO
}
\description{
Automatically called by the \code{\link{add}} functions.
}
