% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,ATO_det-method}
\alias{summary,ATO_det-method}
\title{Summary method for an ATO_det object}
\usage{
\S4method{summary}{ATO_det}(object)
}
\arguments{
\item{object}{an ATO_det object}
}
\value{
Nothing. Prints a summary.
}
\description{
Summary method for an ATO_det object
}
