% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_type_methods.R
\name{table_type<-}
\alias{table_type<-}
\alias{table_type<-,ATO-method}
\title{Assign a new table type to the ATO}
\usage{
table_type(x) <- value

\S4method{table_type}{ATO}(x) <- value
}
\arguments{
\item{x}{an \code{\link{ATO}}}

\item{value}{The new table type to be used inside the ATO.}
}
\value{
Nothing, acts directly on x
}
\description{
Assign a new table type to the ATO
}
