% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_type_methods.R
\name{table_type}
\alias{table_type}
\alias{table_type,ATO-method}
\title{What is the table type of this ATO?}
\usage{
table_type(x, expect)

\S4method{table_type}{ATO}(x, expect)
}
\arguments{
\item{x}{an \code{\link{ATO}}}

\item{expect}{An expected type of table.
Errors if the ATO is not of this type.}
}
\value{
if expect is missing, returns the type of table used by the ATO.
If expect is provided, either errors or returns nothing.
}
\description{
What is the table type of this ATO?
}
