% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzone_methods.R
\name{tzone<-}
\alias{tzone<-}
\alias{tzone<-,ATO-method}
\title{Modify the time zone of the ATO}
\usage{
tzone(x) <- value

\S4method{tzone}{ATO}(x) <- value
}
\arguments{
\item{x}{an \code{\link{ATO}}}

\item{value}{the desired time zone}
}
\value{
Nothing, acts directly on x
}
\description{
Modify the time zone of the ATO
}
\examples{
# check the timezone of the example ATO
tzone(example_ato)
summary(get_det(example_ato))

# change the tz to UTC
tzone(example_ato) <- "UTC"
tzone(example_ato)
summary(get_det(example_ato))

# clean_up
tzone(example_ato) <- "Europe/Copenhagen"

}
