% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzone_methods.R
\name{tzone}
\alias{tzone}
\alias{tzone,ATO-method}
\title{What is the timezone of this ATO?}
\usage{
tzone(x)

\S4method{tzone}{ATO}(x)
}
\arguments{
\item{x}{an \code{\link{ATO}}}
}
\value{
Returns the timezone of the ATO.
}
\description{
What is the timezone of this ATO?
}
\examples{
# check the timezone of the example ATO
tzone(example_ato)

}
